Added file processing classes

+++ src/main/java/org/sdw/ingestion/IngestionConfig.java
@@ -28,7 +28,7 @@ import org.apache.commons.configuration2.ex.ConfigurationException;
  */
 public class IngestionConfig extends ConfigReader
 {
-	public final int maxFileSize;
+	public final long maxFileSize;
 	public final String commonRdfFormat;
 	public final Configuration config;
 	public final String[] datasetPaths;
@@ -55,7 +55,7 @@ public class IngestionConfig extends ConfigReader
 		commonRdfFormat = config.getString("commonRdfFormat");
 		hashFunction = config.getString("hashFunction");
 		datasetHashing = config.getString("datasetHashing");
-		maxFileSize = config.getInt("maxFileSize");
+		maxFileSize = config.getLong("maxFileSize");
 		hashFile = config.getString("hashFile");
 		jenaTDBDatabase = config.getString("jenaTDBDatabase");
 		queryDumpDirectory = config.getString("queryDumpDirectory");
